update uninstall script

+++ uninstall.sh
@@ -1,9 +1,20 @@
-#/bin/bash
+#!/bin/bash
 
 echo "Starting to uninstall"
 
-# docker-compose down -v
+set -e
+
+# Stop and remove containers, networks, etc.
 ./dockerclearall.sh
-sudo rm -rf data artifactory*
+
+# Normalize input to lowercase and compare
+input=$(echo "$1" | tr '[:upper:]' '[:lower:]')
+
+if [[ "$input" == "y" || "$input" == "yes" ]]; then
+    sudo rm -rf data artifactory*
+    echo "Volumes removed"
+else
+    echo "Finished and not removing volumes"
+fi
 
 echo "Done"
